Usuario com permissão

+++ src/main/java/br/edu/utfpr/pb/trabalhofinal/service/UsuarioService.java
@@ -2,6 +2,6 @@ package br.edu.utfpr.pb.trabalhofinal.service;
 
 import br.edu.utfpr.pb.trabalhofinal.model.Usuario;
 
-public interface UsuarioService extends CrudService<Usuario, Integer> {
+public interface UsuarioService extends CrudService<Usuario, Long> {
 
 }
